Added custom support for CircleCI Harbormaster build target and callback. Code based on work from phacility.
A custom implementation was created because phacility was waiting for circleci to fix a bug that prevented the circleci test runner from pulling down code from a git tag (Phabricator uploads code to test to a tag as it's "Staging Area") This implementation based on phacility's temporarily abandoned circleci work uses the github api to copy the tag to a branch, and then has circleci use that branch instead of the tag.

Once circle fixes their issue, or phacility finds a better work around, this code will be obsolete.

Note, this code also sends the branch name (if available) of the built target to CircleCI as an env variable.